nxp:mimxrt1050: Add support for MIMXRT1050_EVK
Add configration for MIMXRT1050_EVK with external QSPI flash.
With the Soc security modules like HAB, BEE, DCP and others,
the physical external flash can be treated as internal flash
with the same security feature.

The configration of MIMXRT1050_EVK supports:
	- 8MB QSPI flash in default
	- mbed-bootloader-rt1050-qspi-internal_flash-no_rot-94579e9.bin
	- Ethernet connecing with Pelion
	- FLASHIAP
	- kvstore
	- OTA

For the Hyper flash, please delete the line:
	"target.macros_remove" : ["HYPERFLASH_BOOT"]
and adjust the other settings to match the Hyper flash size and
flash map.

Signed-off-by: Gavin Liu <gang.liu@nxp.com>